# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /v1/person:
    put:
      tags:
        - person-endpoint
      summary: Create new or update existing person
      description: >-
        Creates new or updates existing person. Returns created/updated person
        with id.
      operationId: add
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            maxLength: 40
            minLength: 8
            type: string
        - name: token
          in: header
          required: false
          schema:
            maxLength: 40
            minLength: 2
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Person'
          application/xml:
            schema:
              $ref: '#/components/schemas/Person'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
            application/xml:
              schema:
                $ref: '#/components/schemas/Person'
  /v1/persons:
    get:
      tags:
        - person-endpoint
      summary: Get all persons
      description: >-
        Returns first N persons specified by the size parameter with page offset
        specified by page parameter.
      operationId: getAll
      parameters:
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            format: int32
        - name: page
          in: query
          description: Zero-based page index
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePerson'
            application/xml:
              schema:
                $ref: '#/components/schemas/PagePerson'
  /v1/person/{id}:
    get:
      tags:
        - person-endpoint
      summary: Verify the person
      description: Verifies the given person and sends an email to his manager, if found guilty.
      operationId: get
      parameters:
        - name: id
          in: path
          description: Person id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Person not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
            application/xml:
              schema:
                $ref: '#/components/schemas/Person'
components:
  schemas:
    Address:
      required:
        - city
        - line1
        - state
        - zip
      type: object
      properties:
        id:
          type: integer
          format: int64
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
    Person:
      required:
        - firstName
        - lastName
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
          enum:
            - M
            - F
        addresses:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Address'
    PagePerson:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Person'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean